Refactoring stylecomponents into smaller files
