+++ frontend/src/styles/stylecomponents/StyledComponentsLibrary.js
@@ -1,181 +1,15 @@
 import styled from "styled-components"
 
-export const Header = styled.div`
-display: flex;
-justify-content: space-between;
-  background-color: ${({ theme }) => theme.colors.backgroundAlt};
-  top: 0;
-`
 
-// Container komponenter
-export const Container = styled.div`
-  width: 100%;
-  min-height: 90vh;
-  background-color: ${({ theme }) => theme.colors.backgroundAlt};
-  display: grid;
-  
-  place-items: center;
-  //overflow: hidden;
-  perspective: 100rem;
-`
-export const PageContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px;
-  max-width: ${({ maxWidth }) => maxWidth || "1200px"};
-  margin: ${({ centered }) => centered ? "auto" : "0"};
-`
 
-// Button komponenter
-export const PrimaryButton = styled.button`
-  padding: ${({ theme }) => theme.components.button.padding};
-  border: ${({ theme }) => theme.components.button.border};
-  border-radius: ${({ theme }) => theme.components.button.borderRadius};
-  background-color: ${({ theme }) => theme.colors.primary};
-  color:${({ theme }) => theme.colors.text} ;
-  cursor: pointer;
-  transition: ${({ theme }) => theme.components.button.transition};
-  font-family: ${({ theme }) => theme.fonts.body};
-  margin: ${({ theme }) => theme.components.button.margin};
 
-  &:hover {
-    background-color: ${({ theme }) => theme.colors.primaryHover};
-  }
 
-  &:disabled {
-    opacity: 0.6;
-    cursor: not-allowed;
-  }
-`
 
-export const ButtonGroup = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 10px;
-  margin-top: 20px;
 
-  @media (min-width: ${({ theme }) => theme.breakpoints.tablet}) {
-    flex-direction: row;
-    justify-content: center;
-  }
-`
 
-// Input komponenter
-export const StyledInput = styled.input`
-  width: 100%;
-  padding: ${({ theme }) => theme.components.input.padding};
-  margin: ${({ theme }) => theme.components.input.margin};
-  border: ${({ theme }) => theme.components.input.border};
-  border-radius: ${({ theme }) => theme.components.input.borderRadius};
-  font-family: ${({ theme }) => theme.fonts.body};
-
-  &:focus {
-    outline: none;
-    border-color: ${({ theme }) => theme.colors.primary};
-  }
-`
-
-// Grid komponenter
-export const GridLayout = styled.div`
-  display: grid;
-  grid-template-columns: 1fr;
-  gap: ${({ theme }) => theme.spacing(2)};
-  align-items: center;
-  justify-items: center;
-  align-content: center;
-  justify-content: center;
-  place-items: center;
-  padding: ${({ theme }) => theme.spacing(2)};
-
-  @media (min-width: ${({ theme }) => theme.breakpoints.tablet}) {
-    grid-template-columns: repeat(2, 1fr);
-  }
-
-  @media (min-width: ${({ theme }) => theme.breakpoints.desktop}) {
-    grid-template-columns: repeat(4, 1fr);
-  }
-`
-
-export const FooterIcon = styled.div`
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 24px;
-  color: ${({ theme }) => theme.colors.text};
-  cursor: pointer;
-  margin: 0 10px;
-  transition: color 0.3s ease;
-  
-  &:hover {
-    color: ${({ theme }) => theme.colors.primary};
-  }
-
-  &.disabled {
-    opacity: 0.5;
-    cursor: not-allowed;
-  }
-
-  /* Tvinga React Icons storlek */
-  svg {
-    width: 24px !important;
-    height: 24px !important;
-  }
-`
-
-// Footer komponenter
-export const FooterStyled = styled.footer`
-position: fixed;
-  bottom: 0;
-  width: 100%;
-  background-color: ${({ theme }) => theme.colors.backgroundAlt};
-  text-align: center;
-  padding: 10px;
-  box-shadow: 0 -2px 5px ${({ theme }) => theme.colors.shadow};
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: ${({ theme }) => theme.spacing(2)};
-`
-
-// Search komponenter
-export const SearchBar = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 20px;
-  gap: 10px;
-  width: 100%;
-  max-width: 600px;
-`
 
 // PlantList specifika komponenter
-export const PlantListContainer = styled.div`
-  padding: ${({ theme }) => theme.spacing(2.5)};
-  max-width: 1200px;
-  margin: 0 auto;
-  min-height: 60vh;
 
-    /* If you want the plant list itself to be scrollable with fixed height: */
-  
-  max-height: 60vh;
-  overflow-y: auto;
-  
-  &::-webkit-scrollbar {
-    width: 8px;
-  }
-
-  &::-webkit-scrollbar-track {
-    background: ${({ theme }) => theme.colors.backgroundAlt};
-    border-radius: 4px;
-  }
-
-  &::-webkit-scrollbar-thumb {
-    background: ${({ theme }) => theme.colors.primary};
-    border-radius: 4px;
-  }
-`
 
 export const ErrorMessage = styled.div`
   color: #e74c3c;
@@ -189,18 +23,6 @@ export const ErrorMessage = styled.div`
 
 `
 
-export const PlantsGrid = styled.div`
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(280px, 1fr));
-  gap: 1rem;
-  padding: 1rem;
-  width: 100%;
-`
-
-export const ThemedPlantsGrid = styled(PlantsGrid)`
-  grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
-  gap: ${({ theme }) => theme.spacing(2.5)};
-`
 
 // Card komponenter
 export const PlantCard = styled.div`
@@ -408,78 +230,3 @@ export const StyledP = styled.p`
   }
 `
 
-export const RemoveButton = styled.button`
-  background: none;
-  border: none;
-  color: #dc3545;
-  cursor: pointer;
-  font-size: 1.2rem;
-  padding: 8px;
-  border-radius: 4px;
-  transition: all 0.2s ease;
-  margin-top: 10px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  min-width: 40px;
-  min-height: 40px;
-
-  &:hover {
-    background-color: #f8d7da;
-    color: #721c24;
-    transform: scale(1.1);
-  }
-
-  &:active {
-    transform: scale(0.95);
-  }
-
-  &:disabled {
-    opacity: 0.5;
-    cursor: not-allowed;
-    
-    &:hover {
-      background-color: transparent;
-      transform: none;
-    }
-  }
-
-  &:focus {
-    outline: 2px solid #dc3545;
-    outline-offset: 2px;
-  }
-
-  /* Alternativ med text */
-  &.with-text {
-    flex-direction: row;
-    gap: 8px;
-    padding: 10px 16px;
-    font-size: 0.9rem;
-    
-    &::after {
-      content: "Ta bort";
-    }
-  }
-
-  /* Varning-stil för viktigare borttagningar */
-  &.warning {
-    color: #856404;
-    
-    &:hover {
-      background-color: #fff3cd;
-      color: #533f03;
-    }
-  }
-
-  /* Minimal stil */
-  &.minimal {
-    padding: 4px;
-    font-size: 1rem;
-    min-width: 30px;
-    min-height: 30px;
-    
-    &:hover {
-      background-color: rgba(220, 53, 69, 0.1);
-    }
-  }
-`
